requirements.txt
pandas==2.2.0
numpy==1.26.4
yfinance>=1.2.0
oandapyV20==0.7.2
boto3==1.34.0
python-dotenv==1.0.0
ta==0.11.0
requests==2.31.0
schedule==1.2.1
pytest==8.0.0
